third_party/blink/web_tests/external/wpt/trusted-types/trusted-types-svg-script-set-href-expected.txt

This is a testharness.js-based test.
[FAIL] Assign string to SVGScriptElement.href.baseVal.
  assert_equals: expected "SVGScriptElement href|about:blank" but got "SVGAnimatedString baseVal|about:blank"
[FAIL] Assign string to SVGScriptElement.className.baseVal.
  Failed to set the 'baseVal' property on 'SVGAnimatedString': This document requires 'TrustedScriptURL' assignment.
[FAIL] Assign string to SVGScriptElement.href.baseVal  w/ default policy.
  assert_equals: expected "SVGScriptElement href" but got "SVGAnimatedString baseVal"
[FAIL] Assign string to SVGScriptElement.className.baseVal w/ default policy.
  assert_equals: expected "SVGScriptElement href" but got "SVGAnimatedString baseVal"
Harness: the test ran to completion.
